basic_strands_agent/requirements.txt
strands-agents>=0.1.0,<0.2.0
strands-agents-tools>=0.1.0,<0.2.0
boto3>=1.26.0,<2.0.0
python-dotenv>=1.0.0
fastapi>=0.95.0,<1.0.0
uvicorn>=0.21.0,<1.0.0
